Move lp_diff to a more convenient location
