+++ pyomo/repn/tests/lp_diff.py
